Move testutil package into util package
No functional changes
