+++ pkg/composition/BUILD.bazel
@@ -86,8 +86,8 @@ go_test(
         "//pkg/k8s/updater:go_default_library",
         "//pkg/options:go_default_library",
         "//pkg/synchronization/api:go_default_library",
-        "//pkg/testutil:go_default_library",
         "//pkg/util:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/ash2k/stager:go_default_library",
         "//vendor/github.com/atlassian/ctrl:go_default_library",
         "//vendor/github.com/atlassian/ctrl/apis/condition/v1:go_default_library",
