+++ pkg/composition/plumbing_for_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/atlassian/voyager"
 	comp_v1 "github.com/atlassian/voyager/pkg/apis/composition/v1"
-	"github.com/atlassian/voyager/pkg/testutil"
+	"github.com/atlassian/voyager/pkg/util/testutil"
 	"github.com/ghodss/yaml"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
