+++ pkg/creator/BUILD.bazel
@@ -50,10 +50,10 @@ go_test(
         "//pkg/creator/ssam:go_default_library",
         "//pkg/pagerduty:go_default_library",
         "//pkg/servicecentral:go_default_library",
-        "//pkg/testutil:go_default_library",
         "//pkg/util/auth:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/mock:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
