+++ pkg/creator/ssam/BUILD.bazel
@@ -34,10 +34,10 @@ go_test(
     embed = [":go_default_library"],
     race = "on",
     deps = [
-        "//pkg/testutil:go_default_library",
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/pkiutil/testutil:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
     ],
