+++ pkg/creator/ssam/it/BUILD.bazel
@@ -12,9 +12,9 @@ go_test(
     ],
     deps = [
         "//pkg/creator/ssam:go_default_library",
-        "//pkg/testutil:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/pkiutil:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes/scheme:go_default_library",
