+++ pkg/execution/plugins/atlassian/iamrole/BUILD.bazel
@@ -34,7 +34,7 @@ go_test(
     deps = [
         "//:go_default_library",
         "//pkg/orchestration/wiring/wiringutil/oap:go_default_library",
-        "//pkg/testutil:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/apis/smith/v1:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/plugin:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
