+++ pkg/formation/controller_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/atlassian/voyager/pkg/k8s"
 	orchclient_fake "github.com/atlassian/voyager/pkg/orchestration/client/fake"
 	orchInf "github.com/atlassian/voyager/pkg/orchestration/informer"
-	"github.com/atlassian/voyager/pkg/testutil"
 	"github.com/atlassian/voyager/pkg/util"
+	"github.com/atlassian/voyager/pkg/util/testutil"
 	"github.com/ghodss/yaml"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/assert"
