+++ pkg/orchestration/wiring/BUILD.bazel
@@ -38,8 +38,8 @@ go_test(
         "//pkg/orchestration/wiring/legacy:go_default_library",
         "//pkg/orchestration/wiring/registry:go_default_library",
         "//pkg/orchestration/wiring/wiringplugin:go_default_library",
-        "//pkg/testutil:go_default_library",
         "//pkg/util/layers:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/apis/smith/v1:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/plugin:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
