+++ pkg/servicecentral/BUILD.bazel
@@ -42,13 +42,13 @@ go_test(
     deps = [
         "//:go_default_library",
         "//pkg/apis/creator/v1:go_default_library",
-        "//pkg/testutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/auth:go_default_library",
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/pkiutil:go_default_library",
         "//pkg/util/pkiutil/testutil:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/mock:go_default_library",
