+++ pkg/util/httptest/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     importpath = "github.com/atlassian/voyager/pkg/util/httptest",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/testutil:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/google/go-cmp/cmp:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
