+++ pkg/util/pkiutil/BUILD.bazel
@@ -36,7 +36,7 @@ go_test(
     embed = [":go_default_library"],
     race = "on",
     deps = [
-        "//pkg/testutil:go_default_library",
+        "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
